+++ scripts/production.js
@@ -1,1 +1,1 @@
-require('cli').fatal('Running production server has not been implemented yet.');
+require('cli').fatal('Running production server has not been implemented yet. Use `node dist/server.js`');